+++ file/mod.ts
@@ -4,4 +4,6 @@ export type {
   RotatingFileSinkDriver,
   RotatingFileSinkOptions,
 } from "./filesink.base.ts";
+export type { StreamFileSinkOptions } from "./streamfilesink.ts";
 export { getFileSink, getRotatingFileSink } from "#filesink";
+export { getStreamFileSink } from "./streamfilesink.ts";